Added Dictionaries to database and controller

+++ src/NBsoft.Wordzz.Contracts/Entities/BoardTile.cs
@@ -5,8 +5,8 @@ namespace NBsoft.Wordzz.Contracts.Entities
     [Serializable]
     public class BoardTile : IBoardTile
     {
-        public short X { get; set; }
-        public short Y { get; set; }
+        public int X { get; set; }
+        public int Y { get; set; }
         public BonusType Bonus { get; set; }
     }
 }
